nhl-hockey-milstone-3/requirements.txt
python-dotenv
numpy
pandas
matplotlib
seaborn
requests
tqdm
lxml
plotly
setuptools
flask
waitress
opencv-python
scikit-learn
comet_ml
jupyterlab
ipywidgets
xgboost